+++ src/controllers/tasks.ts
@@ -1,6 +1,7 @@
+import { Schema } from 'mongoose'
 import { DtoTasks } from '../dto-interfaces/tasks.dto'
 import { ITasks, TasksModel } from '../models/tasks'
-import { ErrorMessagesForTasksController as EFT } from './errors/errors.messages'
+import { ErrorMessagesForTasksController as EFT, GeneralErrorMessages as GEM } from './errors/errors.messages'
 
 class Tasks {
   private _args: DtoTasks | null
@@ -21,6 +22,8 @@ class Tasks {
         return this._store()
       case 'update':
         return this._update()
+      case 'updateStatus':
+        return this._updateStatus()
       default:
         return undefined
     }
@@ -29,19 +32,21 @@ class Tasks {
   private async _deleteOne (): Promise<ITasks | null> {
     const { id } = this._args as DtoTasks
     try {
-      const deletedTasks = await TasksModel.findOneAndDelete({ id })
+      const deletedTasks = await TasksModel.findByIdAndDelete(id)
 
       return deletedTasks
     } catch (error) {
       console.error(error)
-      throw new Error(EFT.problemDeleting)
+      throw new Error(EFT.problemDeletingTasks)
     }
   }
 
   private async _getAll (): Promise<ITasks[]> {
-    const { idCompany, idProject } = this._args as DtoTasks
+    const { idProject } = this._args as DtoTasks
     try {
-      const tasks = await TasksModel.find({ idCompany, idProject })
+      const tasks = await TasksModel.find({
+        idProject: new Schema.Types.ObjectId(idProject as string)
+      })
 
       return tasks
     } catch (error) {
@@ -52,45 +57,56 @@ class Tasks {
 
   private async _store (): Promise<ITasks> {
     const {
-      idCompany,
       idProject,
-      limitDate,
-      name,
+      deadline,
+      description,
       responsible,
       status,
       subTasks
     } = this._args as DtoTasks
 
-    if (!idCompany) throw new Error(EFT.missingIdCompany)
-    else if (!idProject) throw new Error(EFT.missingIdProject)
-    else if (!limitDate) throw new Error(EFT.missingLimitDate)
-    else if (new Date(limitDate).toString() === 'Invalid Date') throw new Error(EFT.invalidDate)
-    else if (!name) throw new Error(EFT.missingName)
-    else if (!responsible) throw new Error(EFT.missingResponsible)
-    else if (!status) throw new Error(EFT.missingStatus)
-
     try {
-      const newTask = new TasksModel({
-        idCompany,
-        idProject,
-        limitDate: new Date(limitDate),
-        name,
-        responsible,
-        status,
-        subTasks
-      })
+      if (!idProject) throw new Error(GEM.missingIdProject)
+      else if (!deadline) throw new Error(GEM.missingDeadline)
+      else if (new Date(deadline).toString() === 'Invalid Date')
+        throw new Error(GEM.invalidDate)
+      else if (!description) throw new Error(GEM.missingDescription)
+      else if (!responsible) throw new Error(EFT.missingResponsible)
+      else if (!status) throw new Error(GEM.missingStatus)
+      else if (!subTasks) throw new Error(EFT.missingSubTasks)
+
+      let newTask: ITasks
+      if (subTasks && subTasks.length > 0)
+        newTask = new TasksModel({
+          deadline         : new Date(deadline),
+          description,
+          idProject,
+          responsible,
+          responsibleNumber: responsible.length,
+          status,
+          subTasks
+        })
+      else
+        newTask = new TasksModel({
+          deadline         : new Date(deadline),
+          description,
+          idProject,
+          responsible,
+          responsibleNumber: responsible.length,
+          status
+        })
       const result = await newTask.save()
 
       return result
     } catch (error) {
       if (
-        error.message === EFT.missingIdCompany ||
-        error.message === EFT.missingIdProject ||
-        error.message === EFT.missingLimitDate ||
-        error.message === EFT.invalidDate ||
-        error.message === EFT.missingName ||
+        error.message === GEM.missingIdProject ||
+        error.message === GEM.missingDeadline ||
+        error.message === GEM.invalidDate ||
+        error.message === GEM.missingDescription ||
         error.message === EFT.missingResponsible ||
-        error.message === EFT.missingStatus
+        error.message === GEM.missingStatus ||
+        error.message === EFT.missingSubTasks
       )
         throw error
       else {
@@ -103,43 +119,58 @@ class Tasks {
   private async _update (): Promise<ITasks | null> {
     const {
       id,
-      limitDate,
-      name,
+      deadline,
+      description,
       responsible,
       status,
       subTasks
     } = this._args as DtoTasks
 
-    if (!limitDate) throw new Error(EFT.missingLimitDate)
-    else if (new Date(limitDate).toString() === 'Invalid Date') throw new Error(EFT.invalidDate)
-    else if (!name) throw new Error(EFT.missingName)
-    else if (!responsible) throw new Error(EFT.missingResponsible)
-    else if (!status) throw new Error(EFT.missingStatus)
-    else if (!subTasks) throw new Error(EFT.missingSubTasks)
-
     try {
-      const result = await TasksModel.findByIdAndUpdate(
-        id,
-        {
-          limitDate: new Date(limitDate),
-          name,
-          responsible,
-          status,
-          subTasks
-        },
-        {
-          new: true
-        }
-      )
+      if (!deadline) throw new Error(GEM.missingDeadline)
+      else if (new Date(deadline).toString() === 'Invalid Date') throw new Error(GEM.invalidDate)
+      else if (!description) throw new Error(GEM.missingDescription)
+      else if (!responsible) throw new Error(EFT.missingResponsible)
+      else if (!status) throw new Error(GEM.missingStatus)
+      else if (!subTasks) throw new Error(EFT.missingSubTasks)
+
+      let result: ITasks | null
+      if (subTasks.length > 0)
+        result = await TasksModel.findByIdAndUpdate(
+          id,
+          {
+            deadline: new Date(deadline),
+            description,
+            responsible,
+            status,
+            subTasks
+          },
+          {
+            new: true
+          }
+        )
+      else
+        result = await TasksModel.findByIdAndUpdate(
+          id,
+          {
+            deadline: new Date(deadline),
+            description,
+            responsible,
+            status
+          },
+          {
+            new: true
+          }
+        )
 
       return result
     } catch (error) {
       if (
-        error.message === EFT.missingLimitDate ||
-        error.message === EFT.invalidDate ||
-        error.message === EFT.missingName ||
+        error.message === GEM.missingDeadline ||
+        error.message === GEM.invalidDate ||
+        error.message === GEM.missingDescription ||
         error.message === EFT.missingResponsible ||
-        error.message === EFT.missingStatus ||
+        error.message === GEM.missingStatus ||
         error.message === EFT.missingSubTasks
       )
         throw error
@@ -147,6 +178,22 @@ class Tasks {
         console.error(error)
         throw new Error(EFT.problemStoringTasks)
       }
+    }
+  }
+
+  private async _updateStatus (): Promise<ITasks | null> {
+    const { id, status } = this._args as DtoTasks
+
+    try {
+      const result = await TasksModel.findByIdAndUpdate(
+        id as string,
+        { status: status as string }
+      )
+
+      return result
+    } catch (error) {
+      console.error(error)
+      throw error
     }
   }
 }