+++ src/dto-interfaces/utils/informationInterface.ts
@@ -1,0 +1,6 @@
+interface IInformation {
+  content: string
+  title  : string
+}
+
+export { IInformation }