+++ src/models/utils/extendTaskSchema.ts
@@ -1,0 +1,11 @@
+import { Schema } from 'mongoose'
+import { CommonDataForProjectsTasksAndJobOffers } from './commonDataForProjectsTasksAndJobOffers'
+
+const extendSchema = (schemaToExtend: Schema): Schema => {
+  return new Schema({
+    ...CommonDataForProjectsTasksAndJobOffers.obj,
+    ...schemaToExtend.obj
+  })
+}
+
+export { extendSchema }