+++ src/models/utils/information.ts
@@ -1,0 +1,14 @@
+import { Schema } from 'mongoose'
+
+const Information = new Schema({
+  content: {
+    required: true,
+    type    : String
+  },
+  title: {
+    required: true,
+    type    : String
+  }
+})
+
+export { Information }