+++ src/routes/tasks.ts
@@ -4,6 +4,7 @@ import { Request } from '../custom/express.request'
 import { response } from '../network/response'
 import { Tasks as TasksC } from '../controllers/tasks'
 import { DtoTasks } from '../dto-interfaces/tasks.dto'
+import { ITasks } from '../models/tasks'
 
 const Tasks = Router()
 
@@ -16,15 +17,11 @@ Tasks.route('/tasks/:idCompany/:idProject/')
       const result = await tc.process('getAll')
       response(false, { result }, res, 200)
     } catch (error) {
-      console.error(error)
       response(true, { message: error.message }, res, 500)
     }
   })
   .post(async (req: Request, res: Response): Promise<void> => {
-    const {
-      body: { args },
-      params: { idCompany, idProject }
-    } = req
+    const { body: { args }, params: { idCompany, idProject } } = req
     const tc = new TasksC({
       ...args as DtoTasks,
       idCompany,
@@ -35,27 +32,46 @@ Tasks.route('/tasks/:idCompany/:idProject/')
       const result = await tc.process('store')
       response(false, { result }, res, 200)
     } catch (error) {
-      console.error(error)
       response(true, { message: error.message }, res, 500)
     }
   })
 
-Tasks.route('/tasks/:idCompany/:idProject/:idTask/')
+Tasks.route('/tasks/:idProject/:idTask/')
   .patch(async (req: Request, res: Response): Promise<void> => {
-    const { body: { args }, params: { idCompany, idProject, idTask } }= req
-    const tc = new TasksC({
-      ...args as DtoTasks,
-      id: idTask,
-      idCompany,
-      idProject
-    } as DtoTasks)
+    const {
+      body: { args },
+      params: { idProject, idTask },
+      query
+    } = req
+    let tc    : TasksC
+    let result: ITasks | ITasks[] | null | undefined
 
-    try {
-      const result = await tc.process('update')
-      response(false, { result }, res, 200)
-    } catch (error) {
-      console.error(error)
-      response(true, { message: error.message }, res, 500)
+    if (Object.keys(query).length === 0) {
+      tc = new TasksC({
+        ...args as DtoTasks,
+        id: idTask,
+        idProject
+      } as DtoTasks)
+
+      try {
+        result = await tc.process('update')
+        response(false, { result }, res, 200)
+      } catch (error) {
+        response(true, { message: error.message }, res, 500)
+      }
+    } else {
+      const { status } = query
+      tc = new TasksC({
+        id    : idTask as string,
+        status: status as string
+      } as DtoTasks)
+
+      try {
+        result = await tc.process('updateStatus')
+        response(false, { result }, res, 200)
+      } catch (error) {
+        response(true, { message: error.message }, res, 500)
+      }
     }
   })
 
@@ -68,7 +84,6 @@ Tasks.route('/task/:idTask/')
       const result = await tc.process('deleteOne')
       response(false, { result }, res, 200)
     } catch (error) {
-      console.error(error)
       response(true, { message: error.message }, res, 500)
     }
   })